add allen.js
