+++ index.js
@@ -1,19 +1,1 @@
-var sizeOf = require('image-size');
-
-
-function ratio(imgPath, width){
-    var dimensions = sizeOf(imgPath);
-    return dimensions.width/dimensions.height;
-}
-
-function width(imgPath){
-    var dimensions = sizeOf(imgPath);
-    return dimensions.width;
-}
-
-function height(imgPath){
-    var dimensions = sizeOf(imgPath);
-    return dimensions.height;
-}
-
-module.exports = {ratio, width, height};
+console.log("Hello Yan")
